+++ mplapack/reference/Cgtrfs.cpp
@@ -182,9 +182,9 @@ void Cgtrfs(const char *trans, INTEGER const n, INTEGER const nrhs, COMPLEX *dl,
         s = zero;
         for (i = 1; i <= n; i = i + 1) {
             if (rwork[i - 1] > safe2) {
-                s = max(s, abs1(work[i - 1]) / rwork[i - 1]);
+                s = max(s, REAL(abs1(work[i - 1]) / rwork[i - 1]));
             } else {
-                s = max(s, (abs1(work[i - 1]) + safe1) / (rwork[i - 1] + safe1));
+                s = max(s, REAL((abs1(work[i - 1]) + safe1) / (rwork[i - 1] + safe1)));
             }
         }
         berr[j - 1] = s;
@@ -264,7 +264,7 @@ void Cgtrfs(const char *trans, INTEGER const n, INTEGER const nrhs, COMPLEX *dl,
         //
         lstres = zero;
         for (i = 1; i <= n; i = i + 1) {
-            lstres = max(lstres, abs1(x[(i - 1) + (j - 1) * ldx]));
+            lstres = max(lstres, REAL(abs1(x[(i - 1) + (j - 1) * ldx])));
         }
         if (lstres != zero) {
             ferr[j - 1] = ferr[j - 1] / lstres;
